+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
   `sign`/`signAndBroadcast`/`signAndBroadcastSync` and related functions now
   have an additional parameter to specify the timeout height. After this height,
   a signed transaction will be considered invalid by the chain. ([#1489])
+- @cosmjs/amino: Export `omitDefault` to help build Amino JSON converters.
 
 [#1489]: https://github.com/cosmos/cosmjs/pull/1489
 
@@ -19,8 +20,12 @@ and this project adheres to
 
 - @cosmjs/stargate: Handle key value pairs in tx search correctly if the value
   is a numeric value. ([#1462])
+- @cosmjs/cosmwasm-stargate: Make `fix_msg` optional in
+  `AminoMsgInstantiateContract2` and omit default in the Amino JSON converter to
+  fix Amino JSON signing for MsgInstantiateContract2. ([#1509])
 
 [#1462]: https://github.com/cosmos/cosmjs/issues/1462
+[#1509]: https://github.com/cosmos/cosmjs/pull/1509
 
 ### Changed